split desmos common code into packages folder added scripts for publishing and versioning on crates.io completed circle CI new workflow
