+++ Cargo.toml
@@ -1,0 +1,17 @@
+[workspace]
+members = ["packages/*", "contracts/*"]
+
+[profile.release.package.desmos]
+codegen-units = 1
+incremental = false
+
+[profile.release]
+opt-level = 3
+debug = false
+rpath = false
+lto = true
+debug-assertions = false
+codegen-units = 1
+panic = 'abort'
+incremental = false
+overflow-checks = true
