+++ contracts/filter-posts/README.md
@@ -1,0 +1,9 @@
+# Filter posts contract
+This contract will filter the posts returning only the posts that has less than `n` reports.
+`n` will be the number of reports that the contract need to exclude a post from the resulting list.
+The parameter can be chosen by the system, or the user that performs the query.
+
+![filter-posts-contract-flow](docs/cw1-filter-flow.png)
+
+## Upload and Interact with the contract on Desmos chain
+[TODO]
