+++ contracts/filter-posts/deploy/desmos_contracts.wasm
