+++ contracts/filter-posts/helper.ts
