+++ contracts/filter-posts/schema/init_msg.json
