+++ contracts/filter-posts/schema/posts_query_response.json
