+++ contracts/filter-posts/schema/query_msg.json
