+++ contracts/filter-posts/src/error.rs
