+++ contracts/filter-posts/src/msg.rs
