+++ contracts/filter-posts/src/state.rs
