+++ packages/desmos/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "desmos"
+version = "0.1.0"
+authors = ["bragaz <leo.braga95@gmail.com>"]
+edition = "2018"
+description = "Desmos network types and helpers"
+license = "Apache 2.0"
+repository = "https://github.com/desmos-labs/desmos-contracts"
+
+[dependencies]
+cosmwasm-std = { version = "0.13.2" }
+schemars = "0.7"
+serde = { version = "1.0.103", default-features = false, features = ["derive"] }
+
+[dev-dependencies]
+cosmwasm-schema = { version = "0.13.2" }
