+++ packages/desmos/examples/schema.rs
@@ -3,9 +3,7 @@ use std::fs::create_dir_all;
 
 use cosmwasm_schema::{export_schema, remove_schemas, schema_for};
 
-use desmos_contracts::custom_query::{PostsQueryResponse, ReportsQueryResponse};
-use desmos_contracts::msg::{HandleMsg, InitMsg, QueryMsg};
-use desmos_contracts::state::State;
+use custom_query::{PostsQueryResponse, ReportsQueryResponse};
 use desmos_contracts::types::{Attachment, OptionalData, PollAnswer, PollData, Post, Report};
 
 fn main() {
@@ -14,10 +12,6 @@ fn main() {
     create_dir_all(&out_dir).unwrap();
     remove_schemas(&out_dir).unwrap();
 
-    export_schema(&schema_for!(InitMsg), &out_dir);
-    export_schema(&schema_for!(HandleMsg), &out_dir);
-    export_schema(&schema_for!(QueryMsg), &out_dir);
-    export_schema(&schema_for!(State), &out_dir);
     export_schema(&schema_for!(OptionalData), &out_dir);
     export_schema(&schema_for!(Attachment), &out_dir);
     export_schema(&schema_for!(PollData), &out_dir);
